Add methods to calculate child benefit and tax
Add methods to calculate the amount of benefit received and tax to
pay.

There is a special case if the tax year selected is `2012`, and
has a different `start_date` which is stored as a constant variable
`TAX_COMMENCEMENT_DATE`.

We have added the child rates to the .yml file which is how rates
are stored for other smart answers and can be accessed using
`@child_benefit_data`.

Co-authored-by: Deborah Chua <deborah.chua@digital.cabinet-office.gov.uk>