Use hint method for money questions
Use the hint method in question presenter for money questions rather 
than the hint_text method. 

The hint_text method was used to prevent the body and hint being split 
across the question title. Since this was issue was identifed with radio 
buttons we are trying to the use the hint method instead which allows 
use to render bullet points and links in the hint text